refactor mcp input analysis modules
